requirements.txt
streamlit
ultralytics
pdf2image
pillow
numpy
opencv-python
qrdet


